\name{CardinalCore-package}
\alias{CardinalCore-package}
\docType{package}

\alias{CardinalCore}

\title{Short description of package}

\description{
Longer description of package.
}

\details{
Detailed description of package.

See \code{browseVignettes("CardinalCore")} to view a user guide and vignettes of common workflows.
}

\author{
Kylie A. Bemis

Maintainer: Kylie A. Bemis <k.bemis@northeastern.edu>
}

\keyword{package}
